backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pymongo==4.6.0
cryptography==41.0.7
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic-core==2.14.1
python-dotenv==1.0.0
cryptography==41.0.7
transformers>=4.38.0
torch>=2.1.0
ctransformers==0.2.0
langdetect==1.0.9
deep-translator==1.8.0
reportlab>=4.0.4
numpy<2
email-validator==2.3.0
